HDPI-7316: address pr comments
